Fix behavior of radio buttons in IE6/7+
IE does not allow the "name" attribute to be set on dynamically created elements.  The workaround is to call createElement() with a non-standard argument: the textual representation of the element with the name attribute already specified, such as '<input name="foo">'.

Without the "name" attribute properly set on the radio input elements, the radio buttons would refuse to be selected.